refactor Party, Characters
