+++ mags/SWNUtilities/GameApp/forms/__init__.py
@@ -1,2 +1,3 @@
 from .ModelChoiceForm import ModelChoiceForm
 from .CharacterDetail import CharacterDetail
+from .AttributeForm import AttributeForm
