+++ mags/SWNUtilities/GameApp/views/Character/CharacterCreateView.py
@@ -1,0 +1,13 @@
+from django.shortcuts import render, redirect
+from django.views.generic.edit import CreateView
+from GameApp.models import Character
+
+
+class CharacterCreate(CreateView):
+    model = Character
+    fields = ['name']
+    template_name = "Form.html"
+
+    def form_valid(self, form):
+        character = form.save()  # save form
+        return redirect('character_update', pk=character.pk)
