+++ mags/SWNUtilities/GameApp/views/Character/CharacterListView.py
@@ -1,0 +1,17 @@
+from django.urls import reverse_lazy
+from django.shortcuts import render, redirect
+from django.views.generic.list import ListView
+from django.views.generic.detail import DetailView
+from django.views.generic.edit import CreateView, UpdateView, DeleteView
+from GameApp.forms import CharacterDetail
+from GameApp.models import Character
+
+
+class CharacterList(ListView):
+    model = Character
+    template_name = "character_list.html"
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['model'] = 'Characters'
+        return context
