+++ mags/SWNUtilities/GameApp/views/Character/__init__.py
@@ -1,0 +1,5 @@
+from .CharacterListView import CharacterList
+from .CharacterDetailView import CharacterDetail
+from .CharacterCreateView import CharacterCreate
+from .CharacterUpdate import CharacterUpdate
+from .CharacterDelete import CharacterDelete
