+++ mags/SWNUtilities/GameApp/views/__init__.py
@@ -1,5 +1,7 @@
 from . import index
-from . import ModelsDropDown
-from .CharacterViews import CharacterList, CharacterCreate, CharacterDetail, CharacterUpdate, CharacterDelete
-from .PartyView import PartyCreate, PartyDetail, PartyList, PartyUpdate, PartyCharactersList
-from .AddCharacterToParty import CharacterChoice
+from .models import ModelsDropDown
+from GameApp.views.Character import CharacterList, CharacterCreate, CharacterDetail, CharacterUpdate, CharacterDelete
+from GameApp.views.Party import PartyCreate, PartyDetail, PartyList, PartyUpdate, PartyDelete, PartyCharacterList
+from GameApp.views.Party import character_choice
+from GameApp.views.Attributes import AttributeCreate
+from GameApp.views.Player import PlayerCreate, PlayerDelete, PlayerList
